requirements.txt
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
comm==0.2.2
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
ipython==8.18.1
ipywidgets==8.1.5
jedi==0.19.2
jupyterlab_widgets==3.0.13
lightgbm==4.6.0
lxml==5.3.1
matplotlib-inline==0.1.7
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
playwright==1.51.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pyee==12.1.1
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
widgetsnbextension==4.0.13
wsproto==1.2.0
